backend/requirements.txt
alembic          
blinker          
Brotli          
cachelib          
certifi          
click          
colorama          
distlib          
filelock          
Flask          
Flask-Caching          
Flask-Compress          
Flask-Cors          
Flask-Migrate          
Flask-SQLAlchemy          
greenlet          
gunicorn          
itsdangerous          
Jinja2          
Mako          
MarkupSafe          
pipenv          
platformdirs          
psycopg2          
python-dotenv          
six          
SQLAlchemy          
typing_extensions          
virtualenv          
virtualenv-clone          
Werkzeug          
